+++ odoo/local-src/sf_rma_sale_order/models/res_company.py
@@ -17,3 +17,5 @@ class ResCompany(models.Model):
 
     rma_service_stockable_product_id = fields.Many2one(
         'product.product')
+
+    rma_service_additional_description = fields.Char(translate=True)