+++ odoo/local-src/sf_rma_sale_order/models/rma_config_settings.py
@@ -31,3 +31,10 @@ class RMASettings(models.TransientModel):
         domain=[('type', '=', 'service')],
         help="This product will appear on RMA sale orders as a placeholder "
              "for imported repair lines with a stockable product.")
+
+    rma_service_additional_description = fields.Char(
+        string='Additional description', translate=True,
+        related='company_id.rma_service_additional_description',
+        help="This text will be appended to the sale order line description, "
+             "after the replaced product name, for imported repair lines."
+    )